sx126x: move RadioController into separate file for clarity
